1. User model refactored.
